+++ tests/fixtures/mv2-basic/worker.js
@@ -1,1 +1,0 @@
-import('./log').then((x) => x.log('hi from worker'))