+++ tests/fixtures/mv3-basic-no-dynamic-import/content.js
@@ -1,2 +1,0 @@
-import { x } from './x.js'
-console.log(x)