+++ tests/fixtures/mv3-both/background.js
@@ -1,7 +1,0 @@
-setTimeout(async () => {
-  const { log } = await import('./log')
-  log('this is background script')
-  chrome.runtime.onMessage.addListener((message) => {
-    log(`receive message from content script`, message)
-  })
-}, 200)