+++ tests/snapshot/mv2-basic/content.js
@@ -1,5 +1,44 @@
 /******/ (() => { // webpackBootstrap
-/******/ 	var __webpack_modules__ = ({});
+/******/ 	"use strict";
+/******/ 	var __webpack_modules__ = ({
+
+/***/ "./test.txt":
+/*!******************!*\
+  !*** ./test.txt ***!
+  \******************/
+/***/ ((module, __unused_webpack_exports, __webpack_require__) => {
+
+module.exports = __webpack_require__.p + "ae771fd2ba5e0558da2f.txt";
+
+/***/ }),
+
+/***/ "./util.js":
+/*!*****************!*\
+  !*** ./util.js ***!
+  \*****************/
+/***/ ((__unused_webpack___webpack_module__, __webpack_exports__, __webpack_require__) => {
+
+__webpack_require__.r(__webpack_exports__);
+/* harmony export */ __webpack_require__.d(__webpack_exports__, {
+/* harmony export */   log: () => (/* binding */ log),
+/* harmony export */   test: () => (/* binding */ test)
+/* harmony export */ });
+function log(label, f) {
+  return async () => {
+    console.group(label)
+    await Promise.resolve().then(f).catch(console.error)
+    console.groupEnd()
+  }
+}
+function test(expr, ...args) {
+  if (expr) console.log('[✅]', ...args)
+  else console.error('[❌]', ...args)
+}
+
+
+/***/ })
+
+/******/ 	});
 /************************************************************************/
 /******/ 	// The module cache
 /******/ 	var __webpack_module_cache__ = {};
@@ -29,6 +68,30 @@
 /******/ 	__webpack_require__.m = __webpack_modules__;
 /******/ 	
 /************************************************************************/
+/******/ 	/* webpack/runtime/WebExtensionBrowserRuntime */
+/******/ 	(() => {
+/******/ 		let isChrome, runtime;
+/******/ 		try {
+/******/ 			if (typeof browser !== "undefined" && typeof browser.runtime?.getURL === "function") {
+/******/ 				runtime = browser;
+/******/ 			}
+/******/ 		} catch (_) {}
+/******/ 		if (!runtime) {
+/******/ 			try {
+/******/ 				if (typeof chrome !== "undefined" && typeof chrome.runtime?.getURL === "function") {
+/******/ 					isChrome = true;
+/******/ 					runtime = chrome;
+/******/ 				}
+/******/ 			} catch (_) {}
+/******/ 		}
+/******/ 		__webpack_require__.webExtRtModern = !isChrome;
+/******/ 		__webpack_require__.webExtRt = runtime || {
+/******/ 			get runtime() {
+/******/ 				throw new Error("No chrome or browser runtime found");
+/******/ 			}
+/******/ 		}
+/******/ 	})();
+/******/ 	
 /******/ 	/* webpack/runtime/define property getters */
 /******/ 	(() => {
 /******/ 		// define getter functions for harmony exports
@@ -59,7 +122,7 @@
 /******/ 		// This function allow to reference async chunks
 /******/ 		__webpack_require__.u = (chunkId) => {
 /******/ 			// return url for filenames based on template
-/******/ 			return "" + chunkId + ".js";
+/******/ 			return "chunks-" + "ed0dde4f184158f0d66f" + ".js";
 /******/ 		};
 /******/ 	})();
 /******/ 	
@@ -80,34 +143,6 @@
 /******/ 		__webpack_require__.o = (obj, prop) => (Object.prototype.hasOwnProperty.call(obj, prop))
 /******/ 	})();
 /******/ 	
-/******/ 	/* webpack/runtime/load script */
-/******/ 	(() => {
-/******/ 		var isBrowser = !!(() => { try { return browser.runtime.getURL("/") } catch(e) {} })()
-/******/ 		var isChrome = !!(() => { try { return chrome.runtime.getURL("/") } catch(e) {} })()
-/******/ 		var runtime = isBrowser ? browser : isChrome ? chrome : { get runtime() { throw new Error("No chrome or browser runtime found") } }
-/******/ 		var __send__ = (msg) => {
-/******/ 			if (isBrowser) return runtime.runtime.sendMessage(msg)
-/******/ 			return new Promise(r => runtime.runtime.sendMessage(msg, r))
-/******/ 		}
-/******/ 		var classicLoader = (url, done) => {
-/******/ 			__send__({ type: 'WTW_INJECT', file: url }).then(done, (e) => done(Object.assign(e, { type: 'missing' })))
-/******/ 		}
-/******/ 		var scriptLoader = (url, done) => {
-/******/ 			var script = document.createElement('script')
-/******/ 			script.src = url
-/******/ 			script.onload = done
-/******/ 			script.onerror = done
-/******/ 			document.body.appendChild(script)
-/******/ 		}
-/******/ 		var workerLoader = (url, done) => {
-/******/ 			try { importScripts(url); done() } catch (e) { done(e) }
-/******/ 		}
-/******/ 		var isWorker = typeof importScripts === 'function'
-/******/ 		if (typeof location === 'object' && location.protocol.includes('-extension:')) __webpack_require__.l = isWorker ? workerLoader : scriptLoader
-/******/ 		else if (!isWorker) __webpack_require__.l = classicLoader
-/******/ 		else { throw new TypeError('Unable to determinate the chunk loader: content script + Worker') }
-/******/ 	})();
-/******/ 	
 /******/ 	/* webpack/runtime/make namespace object */
 /******/ 	(() => {
 /******/ 		// define __esModule on exports
@@ -119,21 +154,90 @@
 /******/ 		};
 /******/ 	})();
 /******/ 	
+/******/ 	/* webpack/runtime/load script */
+/******/ 	(() => {
+/******/ 		let bug816121warned, isNotIframe;
+/******/ 		try {
+/******/ 			isNotIframe = typeof window === "object" ? window.top === window : true;
+/******/ 		} catch(e) {
+/******/ 			isNotIframe = false /* CORS error */;
+/******/ 		}
+/******/ 		const classicLoader = (url, done) => {
+/******/ 			const msg = { type: "WTW_INJECT", file: url };
+/******/ 			const onError = (e) => {
+/******/ 				done(Object.assign(e, { type: "missing" }))
+/******/ 			};
+/******/ 			if (__webpack_require__.webExtRtModern) {
+/******/ 				__webpack_require__.webExtRt.runtime.sendMessage(msg).then(done, onError);
+/******/ 			} else {
+/******/ 				__webpack_require__.webExtRt.runtime.sendMessage(msg, () => {
+/******/ 					const error = __webpack_require__.webExtRt.runtime.lastError;
+/******/ 					if (error) onError(error);
+/******/ 					else done();
+/******/ 				});
+/******/ 			}
+/******/ 		};
+/******/ 		const dynamicImportLoader = (url, done, key, chunkId) => {
+/******/ 			import(url).then(() => {
+/******/ 				if (isNotIframe) return done();
+/******/ 				try {
+/******/ 					// It's a Chrome bug, if the import() is called in a sandboxed iframe, it _fails_ the script loading but _resolve_ the Promise.
+/******/ 					// we call __webpack_require__.f.j(chunkId) to check if it is loaded.
+/******/ 					// if it is, this is a no-op. if it is not, it will throw a TypeError because this function requires 2 parameters.
+/******/ 					// This call will not trigger the chunk loading because it is already loading.
+/******/ 					// see https://github.com/awesome-webextension/webpack-target-webextension/issues/41
+/******/ 					chunkId !== undefined && __webpack_require__.f.j(chunkId);
+/******/ 					done();
+/******/ 				}
+/******/ 				catch (_) {
+/******/ 					if (!bug816121warned) {
+/******/ 						console.warn("Chrome bug https://crbug.com/816121 hit.");
+/******/ 						bug816121warned = true;
+/******/ 					}
+/******/ 					return fallbackLoader(url, done, key, chunkId);
+/******/ 				}
+/******/ 			}, (e) => {
+/******/ 				console.warn("Dynamic import loader failed. Using fallback loader (see https://github.com/awesome-webextension/webpack-target-webextension#content-script).", e);
+/******/ 				fallbackLoader(url, done, key, chunkId);
+/******/ 			});
+/******/ 		}
+/******/ 		const scriptLoader = (url, done) => {
+/******/ 			const script = document.createElement('script');
+/******/ 			script.src = url;
+/******/ 			script.onload = done;
+/******/ 			script.onerror = done;
+/******/ 			document.body.appendChild(script);
+/******/ 		}
+/******/ 		const workerLoader = (url, done) => {
+/******/ 			try {
+/******/ 				importScripts(url);
+/******/ 				done();
+/******/ 			} catch (e) {
+/******/ 				done(e);
+/******/ 			}
+/******/ 		}
+/******/ 		const isWorker = typeof importScripts === 'function'
+/******/ 		if (typeof location === 'object' && location.protocol.includes('-extension:')) {
+/******/ 			__webpack_require__.l = isWorker ? workerLoader : scriptLoader;
+/******/ 		}
+/******/ 		else if (!isWorker) __webpack_require__.l = classicLoader;
+/******/ 		else { throw new TypeError('Unable to determinate the chunk loader: content script + Worker'); }
+/******/ 		const fallbackLoader = __webpack_require__.l;
+/******/ 		__webpack_require__.l = dynamicImportLoader;
+/******/ 	})();
+/******/ 	
 /******/ 	/* webpack/runtime/publicPath */
 /******/ 	(() => {
-/******/ 		var isBrowser = !!(() => { try { return browser.runtime.getURL("/") } catch(e) {} })()
-/******/ 		var isChrome = !!(() => { try { return chrome.runtime.getURL("/") } catch(e) {} })()
-/******/ 		var runtime = isBrowser ? browser : isChrome ? chrome : { get runtime() { throw new Error("No chrome or browser runtime found") } }
-/******/ 		var scriptUrl;
+/******/ 		let scriptUrl;
 /******/ 		if (__webpack_require__.g.importScripts) scriptUrl = __webpack_require__.g.location + "";
-/******/ 		var document = __webpack_require__.g.document;
-/******/ 		if (!scriptUrl && document && document.currentScript) {
-/******/ 			scriptUrl = document.currentScript.src
+/******/ 		const document = __webpack_require__.g.document;
+/******/ 		if (!scriptUrl && document?.currentScript) {
+/******/ 			scriptUrl = document.currentScript.src;
 /******/ 		}
 /******/ 		// When supporting browsers where an automatic publicPath is not supported you must specify an output.publicPath manually via configuration
 /******/ 		// or pass an empty string ("") and set the __webpack_public_path__ variable from your code to use your own logic.
 /******/ 		if (!scriptUrl) {
-/******/ 			if (isChrome || isBrowser) scriptUrl = runtime.runtime.getURL("/");
+/******/ 			if (__webpack_require__.webExtRt) scriptUrl = __webpack_require__.webExtRt.runtime.getURL("/");
 /******/ 			else throw new Error("Automatic publicPath is not supported in this browser");
 /******/ 		}
 /******/ 		scriptUrl = scriptUrl.replace(/#.*$/, "").replace(/\?.*$/, "").replace(/\/[^\/]+$/, "/");
@@ -232,13 +336,45 @@
 /******/ 	
 /************************************************************************/
 var __webpack_exports__ = {};
+// This entry needs to be wrapped in an IIFE because it needs to be isolated against other modules in the chunk.
+(() => {
 /*!********************!*\
   !*** ./content.js ***!
   \********************/
-__webpack_require__.e(/*! import() */ "log_js").then(__webpack_require__.bind(__webpack_require__, /*! ./log */ "./log.js")).then(({ log }) => {
-  log('this is content script')
-  chrome.runtime.sendMessage('Hi!')
-})
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var _util_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./util.js */ "./util.js");
+/// <reference lib="dom" />
+// @ts-check
+
+
+
+Promise.resolve()
+  .then(
+    (0,_util_js__WEBPACK_IMPORTED_MODULE_0__.log)('Test A: import.meta.url', () => {
+      const url = new URL(/* asset import */ __webpack_require__(/*! ./test.txt */ "./test.txt"), __webpack_require__.b).toString()
+      ;(0,_util_js__WEBPACK_IMPORTED_MODULE_0__.test)(url.includes('-extension://'), "new URL('./test.txt', import.meta.url)\n", url)
+    })
+  )
+  .then(
+    (0,_util_js__WEBPACK_IMPORTED_MODULE_0__.log)('Test B: __webpack_public_path__', () => {
+      (0,_util_js__WEBPACK_IMPORTED_MODULE_0__.test)(__webpack_require__.p.includes('-extension://'), '__webpack_public_path__\n', __webpack_require__.p)
+    })
+  )
+  .then(
+    (0,_util_js__WEBPACK_IMPORTED_MODULE_0__.log)('Test C: dynamic import', async () => {
+      console.log("await import('./log.js')\n")
+      const mod = await __webpack_require__.e(/*! import() */ "log_js").then(__webpack_require__.bind(__webpack_require__, /*! ./log.js */ "./log.js"))
+      ;(0,_util_js__WEBPACK_IMPORTED_MODULE_0__.test)('file' in mod, mod)
+    })
+  )
+  .then(() => {
+    chrome.runtime.sendMessage('Hello from content script!')
+    chrome.runtime.onMessage.addListener((message) => {
+      console.log('Message from background:', message)
+    })
+  })
+
+})();
 
 /******/ })()
 ;